refactor: migrate storage to SharedPreferences, update analysis options, and fix linting

+++ lib/features/widget/prayer_widget_service.dart
@@ -254,10 +254,10 @@ class PrayerWidgetService {
     if (kIsWeb) return;
     final settings = await getStoredSettings();
     await updateWidget(
-      latitude: settings['latitude'],
-      longitude: settings['longitude'],
-      method: settings['method'],
-      madhab: settings['madhab'],
+      latitude: settings['latitude'] as double,
+      longitude: settings['longitude'] as double,
+      method: settings['method'] as int,
+      madhab: settings['madhab'] as int,
     );
   }
